requirements.txt
streamlit
boto3
google-auth
google-auth-oauthlib
google-api-python-client
requests
